ONNXRunTime

Documentation for ONNXRunTime.

ONNXRunTime.CreateTensorWithDataAsOrtValue

ONNXRunTime.GetApi

ONNXRunTime.unsafe_GetTensorMutableData

ONNXRunTime.CreateTensorWithDataAsOrtValue — Method

This operation is unsafe, because the return value points to memory owned by data. If data is garbage collected, the return value becomes invalid.

ONNXRunTime.GetApi — Method

Convenience method.

ONNXRunTime.unsafe_GetTensorMutableData — Method

This function is unsafe, because its output points to memory owned by tensor.